WIP: New LoadBalancer API.
RoundRobin and PickFirst load balancers are revamped. Tests yet to be
fixed.

Removed the handleNameResolution*() methods from LoadBalancer, and
replaced it with getNameResolverListener(), so that I can make the
listener not thread-safe, making the LoadBalancer easier to implement.

TODO:
Fix tests.

Transfer the responsibility of shutting down TransportSets completely
from ManagedChannelImpl to LoadBalancer, or fix the race:
  ManagedChannelImpl.shutdown() or going idle
     Save TransportSet collection
       --> Here LoadBalancer may ask for new Subchannel (*)
     LoadBalancer.shutdown()
     Shutdown TransportSet collection
     (If LoadBalancer.shutdown() doesn't shutdown the Subchannels, those
     created at (*) will not be shutdown by anybody).